% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_household_model_excels.R
\name{create_household_model_excels}
\alias{create_household_model_excels}
\title{A wrapper for a Python function to output household projection Excels}
\usage{
create_household_model_excels(
  output_dir,
  wb_filename,
  projection_name,
  model = "both"
)
}
\arguments{
\item{output_dir}{string. The directory from which to read the data and save the Excel file}

\item{wb_filename}{String. The name of the output Excel file. With or without '.xlsx'}

\item{projection_name}{String. The projection name.}

\item{model}{String. Household to output: 'ons', 'dclg' or 'both'. Default 'both'.}
}
\value{
Output Excel workbooks
}
\description{
The Python function read-in model output CSV files and outputs 2 Excel files:
one for the ONS model and one for the DCLG model
}
