% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_model_dclg.R
\name{dclg_stage_1}
\alias{dclg_stage_1}
\title{Implementation of the DCLG 2014 Stage 1 Household Model}
\usage{
dclg_stage_1(population, stage1_file_path)
}
\arguments{
\item{population}{A data frame containing population data.}

\item{stage1_file_path}{String. Path to file containing DCLG
stage 1 household inputs}
}
\value{
A list containing 2 dataframes: Stage 1 household by household
  type and total.
}
\description{
Produce household projections from an input population
projection using the 2014 DCLG household projection methodology
}
