% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/household_model_dclg.R
\name{household_model_dclg}
\alias{household_model_dclg}
\title{Run the 2014 DCLG household model}
\usage{
household_model_dclg(population, stage1_file_path, stage2_file_path)
}
\arguments{
\item{population}{A data frame containing population data.}

\item{stage1_file_path}{String. Path to file containing DCLG
stage 1 household inputs}

\item{stage2_file_path}{String. Path to file containing DCLG
stage 2 household inputs}
}
\value{
A list containing 2 lists: Stage 1 outputs and Stage 2
  outputs.
}
\description{
Run stages 1 and 2 of the DCLG household model and return outputs
}
