% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/output_trend_projection.R
\name{output_projection}
\alias{output_projection}
\title{Output a trend population projection}
\usage{
output_projection(projection, output_dir, write_excel, n_csv_elements)
}
\arguments{
\item{projection}{A list. The model outputs from \code{arrange_trend_core_outputs}}

\item{output_dir}{String. The output location folder path}

\item{write_excel}{Logical. Should the Excel outputs be written}

\item{n_csv_elements}{Numeric. The function will output csv files for elements
1:\code{n_csv_elements} of the \code{projection} list.}
}
\description{
Write csv, rds and xlsx model outputs from the \code{trend_core} cohort component
model. Input is taken from the \code{arrange_trend_core_outputs} function.
}
