% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_trend_model.R
\name{run_trend_model}
\alias{run_trend_model}
\title{Run a trend model config file to produce a population projection
Previously 00_control}
\usage{
run_trend_model(config_list)
}
\arguments{
\item{config_list}{A List. A trend model configuration list.}
}
\description{
Read in, validate and manage model input data and then run the \code{trend_core}
function to produce a population projection. Run the \code{household model} functions
as specified in the config list. Run model output functions and, if specified,
rmarkdown QA scripts.
}
