% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_core.R
\name{trend_core}
\alias{trend_core}
\title{Run a cohort component population model and output population, components of change and rates}
\usage{
trend_core(
  start_population,
  fertility_rates,
  mortality_rates,
  int_out_flows_rates,
  int_in_flows,
  domestic_rates,
  int_out_method,
  constraints = NULL,
  upc = NULL,
  projection_year,
  region_lookup
)
}
\arguments{
\item{start_population}{A data frame. The population at the start of the time period}

\item{fertility_rates}{A data frame. A set of age-secific fertility rates to be applied to the population}

\item{mortality_rates}{A data frame. A set of age/sex-specific mortality probabilities to be
applied to the population}

\item{int_out_flows_rates}{A data frame. A set of international out migration rates to be applied to
the population, or a set of international out flow totals to be subtracted from the population}

\item{int_in_flows}{A data frame. A set of international in migration flows to be added to
the population}

\item{domestic_rates}{A data frame. A set of orgin-destination migration rates by age and sex
to be applied to the population}

\item{int_out_method}{A string. A switch to define whether international out migration is
a rate or a flow. Either \code{rate} or \code{flow}}

\item{constraints}{A list. A set of national-level constraints for each component. If the projection
is to be run unconstrained this is set to NULL. Default \code{NULL}}

\item{upc}{A data frame. Unattributable population change component. If no UPC is being applied this
is set to NULL. Default \code{NULL}}

\item{projection_year}{Numeric. The year being projected}
}
\value{
A list where each element is a data frame containing either projected population or
projected components of change.
}
\description{
Given a starting population and set of projected rates for fertility, mortality and migration
the function will produce a population for the next year using a cohort component method.
The model outputs are the calculated populations and components of change.
}
