% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trend_datastore_outputs.R
\name{trend_datastore_outputs}
\alias{trend_datastore_outputs}
\title{Trend Model Datastore outputs}
\usage{
trend_datastore_outputs(
  population,
  births,
  deaths,
  int_in,
  int_out,
  dom_in,
  dom_out,
  output_dir,
  excel_file_name,
  write_excel
)
}
\arguments{
\item{population, births, deaths, int_in, int_out, dom_in, dom_out}{Data frames with population and component data.}

\item{output_dir}{Path to output directory.}

\item{excel_file_name}{Output file}

\item{write_excel}{Logical. Whether to create Excel output files.}
}
\description{
Trend Model Datastore outputs
}
