% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_houshold_model_excels.R
\name{create_household_model_excels}
\alias{create_household_model_excels}
\title{A wrapper for a Python function to output household projection Excels}
\usage{
create_household_model_excels(output_dir, projection_name)
}
\arguments{
\item{output_dir}{String. The root output directory of the projection}

\item{projection_name}{String. The name of the projection}
}
\value{
Output Excel workbooks
}
\description{
The Python function read-in model output CSV files and outputs 2 Excel files:
one for the ONS model and one for the DCLG model
}
