% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_trend_model_excels.R
\name{create_trend_model_excels}
\alias{create_trend_model_excels}
\title{Produce 3 Excel workbooks from a set of projection RDS files}
\usage{
create_trend_model_excels(output_dir, excel_file_name, household_models = TRUE)
}
\arguments{
\item{output_dir}{String. The output directory where the projection RDS files are
saved.}

\item{excel_file_name}{String. File name for the population workbook output.
With or without "xlsx" suffix.}

\item{household_models}{Logical. Should the household model excel files also
be created.}
}
\description{
If a projection has been run with the \code{write_excel} variable set to
\code{FALSE} and the Excel outputs are subsequently required this function
will produce those workbooks. Creates a population workbook, an ONS household
workbook and a DCLG household workbook.
}
